:recycle: api: move certification-assessment-repository to src
