+++ api/scripts/certification/generate-certification-results.js
@@ -3,7 +3,7 @@ import { knex, disconnect } from '../../db/knex-database-connection.js';
 const ASSESSMENT_COUNT = parseInt(process.env.ASSESSMENT_COUNT) || 100;
 import bluebird from 'bluebird';
 import * as scoringCertificationService from '../../lib/domain/services/scoring/scoring-certification-service.js';
-import * as certificationAssessmentRepository from '../../lib/infrastructure/repositories/certification-assessment-repository.js';
+import * as certificationAssessmentRepository from '../../src/certification/shared/infrastructure/repositories/certification-assessment-repository.js';
 import * as url from 'url';
 
 async function _retrieveLastScoredAssessmentIds() {
