+++ api/scripts/certification/trigger-autojury-on-finalized-sessions.js
@@ -2,7 +2,7 @@ import { knex, disconnect } from '../../db/knex-database-connection.js';
 import bluebird from 'bluebird';
 import { handleAutoJury } from '../../lib/domain/events/handle-auto-jury.js';
 import * as certificationIssueReportRepository from '../../src/certification/shared/infrastructure/repositories/certification-issue-report-repository.js';
-import * as certificationAssessmentRepository from '../../lib/infrastructure/repositories/certification-assessment-repository.js';
+import * as certificationAssessmentRepository from '../../src/certification/shared/infrastructure/repositories/certification-assessment-repository.js';
 import * as certificationCourseRepository from '../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
 import * as challengeRepository from '../../src/shared/infrastructure/repositories/challenge-repository.js';
 import { logger } from '../../lib/infrastructure/logger.js';
